move building logic to AnnotatedPythonPackageBuilder rename the data model analogous to the backend
