+++ client/src/common/MenuBar.tsx
@@ -42,37 +42,13 @@ import {
     resetAnnotations,
     toggleAnnotationImportDialog,
 } from '../features/annotations/annotationSlice';
-import {
-    InferableAnnotation,
-    InferableAttributeAnnotation,
-    InferableBoundaryAnnotation,
-    InferableCalledAfterAnnotation,
-    InferableConstantAnnotation,
-    InferableEnumAnnotation,
-    InferableGroupAnnotation,
-    InferableMoveAnnotation,
-    InferableOptionalAnnotation,
-    InferableRenameAnnotation,
-    InferableRequiredAnnotation,
-    InferableUnusedAnnotation,
-} from '../features/InferableDataModel/InferableAnnotation';
-import InferablePythonClass from '../features/InferableDataModel/InferablePythonClass';
-import InferablePythonFunction from '../features/InferableDataModel/InferablePythonFunction';
-import InferablePythonModule from '../features/InferableDataModel/InferablePythonModule';
-import InferablePythonPackage from '../features/InferableDataModel/InferablePythonPackage';
-import InferablePythonParameter from '../features/InferableDataModel/InferablePythonParameter';
-import InferablePythonResult from '../features/InferableDataModel/InferablePythonResult';
-import PythonClass from '../features/packageData/model/PythonClass';
+import AnnotatedPythonPackageBuilder from '../features/InferableDataModel/AnnotatedPythonPackageBuilder';
 import { PythonFilter } from '../features/packageData/model/PythonFilter';
-import PythonFunction from '../features/packageData/model/PythonFunction';
-import PythonModule from '../features/packageData/model/PythonModule';
 import PythonPackage from '../features/packageData/model/PythonPackage';
 import {
     parsePythonPackageJson,
     PythonPackageJson,
 } from '../features/packageData/model/PythonPackageBuilder';
-import PythonParameter from '../features/packageData/model/PythonParameter';
-import PythonResult from '../features/packageData/model/PythonResult';
 import { togglePackageDataImportDialog } from '../features/packageData/packageDataSlice';
 import { Setter } from './util/types';
 
@@ -185,260 +161,13 @@ const MenuBar: React.FC<MenuBarProps> = function ({ filter, setFilter }) {
         }
     };
 
-    const possibleAnnotations = [
-        'Attribute',
-        'Boundary',
-        'CalledAfters',
-        'Constant',
-        'Enum',
-        'Groups',
-        'Move',
-        'Optional',
-        'Rename',
-        'Required',
-        'Unused',
-    ];
-
-    const getExistingAnnotations = (target: string): InferableAnnotation[] => {
-        let targetAnnotations: InferableAnnotation[] = [];
-        possibleAnnotations.forEach((annotation) => {
-            const returnedAnnotations = returnFormattedAnnotation(
-                target,
-                annotation,
-            );
-            if (returnedAnnotations) {
-                targetAnnotations.concat(returnedAnnotations);
-            }
-        });
-        return targetAnnotations;
-    };
-
-    const returnFormattedAnnotation = (
-        target: string,
-        annotationType: string,
-    ): InferableAnnotation[] | InferableAnnotation | undefined => {
-        switch (annotationType) {
-            case 'Attribute':
-                const attributeAnnotation = annotationStore.attributes[target];
-                if (attributeAnnotation) {
-                    return new InferableAttributeAnnotation(
-                        attributeAnnotation,
-                    );
-                }
-                break;
-            case 'Boundary':
-                const boundaryAnnotation = annotationStore.boundaries[target];
-                if (boundaryAnnotation) {
-                    return new InferableBoundaryAnnotation(boundaryAnnotation);
-                }
-                break;
-            case 'CalledAfters':
-                const calledAfterAnnotations =
-                    annotationStore.calledAfters[target];
-                if (!calledAfterAnnotations) {
-                    break;
-                }
-                const calledAfterAnnotationList = Object.values(
-                    calledAfterAnnotations,
-                );
-                if (calledAfterAnnotationList.length === 1) {
-                    return new InferableCalledAfterAnnotation(
-                        calledAfterAnnotationList[0],
-                    );
-                }
-                let inferableCalledAfterAnnotations: InferableCalledAfterAnnotation[] =
-                    [];
-                calledAfterAnnotationList.forEach((calledAfterAnnotation) => {
-                    inferableCalledAfterAnnotations.push(
-                        new InferableCalledAfterAnnotation(
-                            calledAfterAnnotation,
-                        ),
-                    );
-                });
-                return inferableCalledAfterAnnotations;
-            case 'Constant':
-                const constantAnnotation = annotationStore.constants[target];
-                if (constantAnnotation) {
-                    return new InferableConstantAnnotation(constantAnnotation);
-                }
-                break;
-            case 'Groups':
-                const groupAnnotations = annotationStore.groups[target];
-                if (!groupAnnotations) {
-                    break;
-                }
-                const groupAnnotationList = Object.values(groupAnnotations);
-                if (groupAnnotationList.length === 1) {
-                    return new InferableGroupAnnotation(groupAnnotationList[0]);
-                }
-                let inferableGroupAnnotations: InferableGroupAnnotation[] = [];
-                groupAnnotationList.forEach((groupAnnotation) => {
-                    inferableGroupAnnotations.push(
-                        new InferableGroupAnnotation(groupAnnotation),
-                    );
-                });
-                return inferableGroupAnnotations;
-            case 'Enum':
-                const enumAnnotation = annotationStore.enums[target];
-                if (enumAnnotation) {
-                    return new InferableEnumAnnotation(enumAnnotation);
-                }
-                break;
-            case 'Move':
-                const moveAnnotation = annotationStore.moves[target];
-                if (moveAnnotation) {
-                    return new InferableMoveAnnotation(moveAnnotation);
-                }
-                break;
-            case 'Optional':
-                const optionalAnnotation = annotationStore.optionals[target];
-                if (optionalAnnotation) {
-                    return new InferableOptionalAnnotation(optionalAnnotation);
-                }
-                break;
-            case 'Rename':
-                const renameAnnotation = annotationStore.renamings[target];
-                if (renameAnnotation) {
-                    return new InferableRenameAnnotation(renameAnnotation);
-                }
-                break;
-            case 'Required':
-                const requiredAnnotation = annotationStore.requireds[target];
-                if (requiredAnnotation) {
-                    return new InferableRequiredAnnotation();
-                }
-                break;
-            case 'Unused':
-                const unusedAnnotation = annotationStore.unuseds[target];
-                if (unusedAnnotation) {
-                    return new InferableUnusedAnnotation();
-                }
-                break;
-        }
-        return undefined;
-    };
-
-    const buildInferablePackage = (pythonPackage: PythonPackage) => {
-        const inferablePackage: InferablePythonPackage = {
-            name: pythonPackage.name,
-            distribution: pythonPackage.distribution,
-            version: pythonPackage.version,
-            annotations: getExistingAnnotations(pythonPackage.pathAsString()),
-            modules: buildInferableModules(pythonPackage.modules),
-        };
-        return inferablePackage;
-    };
-
-    const buildInferableClasses = (pythonModule: PythonModule) => {
-        let pythonClasses: InferablePythonClass[] = [];
-        pythonModule.classes.forEach((pythonClass: PythonClass) => {
-            pythonClasses.push({
-                name: pythonClass.name,
-                qualifiedName: pythonClass.qualifiedName,
-                decorators: pythonClass.decorators,
-                superclasses: pythonClass.superclasses,
-                description: pythonClass.description,
-                fullDocstring: pythonClass.fullDocstring,
-                methods: buildInferableFunctions(pythonClass),
-                annotations: getExistingAnnotations(pythonClass.pathAsString()),
-            });
-        });
-        return pythonClasses;
-    };
-
-    const buildInferableFunctions = (
-        pythonDeclaration: PythonModule | PythonClass,
-    ) => {
-        let pythonFunctions: InferablePythonFunction[] = [];
-        if (pythonDeclaration instanceof PythonModule) {
-            pythonDeclaration.functions.forEach(
-                (pythonFunction: PythonFunction) => {
-                    pythonFunctions.push(
-                        buildInferableFunction(pythonFunction),
-                    );
-                },
-            );
-        }
-        if (pythonDeclaration instanceof PythonClass) {
-            pythonDeclaration.methods.forEach(
-                (pythonFunction: PythonFunction) => {
-                    pythonFunctions.push(
-                        buildInferableFunction(pythonFunction),
-                    );
-                },
-            );
-        }
-        return pythonFunctions;
-    };
-
-    const buildInferableFunction = (pythonFunction: PythonFunction) => {
-        return new InferablePythonFunction(
-            pythonFunction.name,
-            pythonFunction.qualifiedName,
-            pythonFunction.decorators,
-            buildInferableParameters(pythonFunction.parameters),
-            buildInferablePythonResults(pythonFunction.results),
-            pythonFunction.isPublic,
-            pythonFunction.description,
-            pythonFunction.fullDocstring,
-            getExistingAnnotations(pythonFunction.pathAsString()),
-        );
-    };
-
-    const buildInferablePythonResults = (pythonResults: PythonResult[]) => {
-        let inferablePythonResults: InferablePythonResult[] = [];
-        pythonResults.forEach((pythonResult: PythonResult) => {
-            inferablePythonResults.push(
-                new InferablePythonResult(
-                    pythonResult.name,
-                    pythonResult.type,
-                    pythonResult.typeInDocs,
-                    pythonResult.description,
-                    getExistingAnnotations(pythonResult.pathAsString()),
-                ),
-            );
-        });
-        return inferablePythonResults;
-    };
-
-    const buildInferableParameters = (pythonParameters: PythonParameter[]) => {
-        let inferablePythonParameters: InferablePythonParameter[] = [];
-        pythonParameters.forEach((pythonParameter: PythonParameter) => {
-            inferablePythonParameters.push(
-                new InferablePythonParameter(
-                    pythonParameter.name,
-                    pythonParameter.defaultValue,
-                    pythonParameter.assignedBy,
-                    pythonParameter.isPublic,
-                    pythonParameter.typeInDocs,
-                    pythonParameter.description,
-                    getExistingAnnotations(pythonParameter.pathAsString()),
-                ),
-            );
-        });
-        return inferablePythonParameters;
-    };
-
-    const buildInferableModules = (pythonModules: PythonModule[]) => {
-        let inferablePythonModules: InferablePythonModule[] = [];
-        pythonModules.forEach((pythonModule: PythonModule) => {
-            inferablePythonModules.push(
-                new InferablePythonModule(
-                    pythonModule.name,
-                    pythonModule.imports,
-                    pythonModule.fromImports,
-                    buildInferableClasses(pythonModule),
-                    buildInferableFunctions(pythonModule),
-                    getExistingAnnotations(pythonModule.pathAsString()),
-                ),
-            );
-        });
-        return inferablePythonModules;
-    };
-
     const infer = () => {
+        const annotatedPythonPackageBuilder = new AnnotatedPythonPackageBuilder(
+            currentPythonPackage,
+            annotationStore,
+        );
         const annotatedPythonPackage =
-            buildInferablePackage(currentPythonPackage);
+            annotatedPythonPackageBuilder.generateAnnotatedPythonPackage();
 
         const requestOptions = {
             method: 'POST',
